Add client part. Extract common server and client parts
