+++ libnet/client.c
@@ -1,0 +1,137 @@
+#include "client.h"
+
+#define _GNU_SOURCE 201112L
+
+#include <arpa/inet.h>
+#include <fdebug.h>
+#include <netdb.h>
+#include <pthread.h>
+#include <sys/select.h>
+#include <sys/socket.h>
+#include <sys/types.h>
+
+#include "common.h"
+
+client_info server[1];
+
+inline unsigned get_client_id(client_info *client) {
+	return 0;
+}
+
+client_info* get_client_by_fd(int fd) {
+	return server;
+}
+
+
+fd_set selects;
+int nfds = 0;
+static void add_fd_to_select(int fd) {
+	log_info("Adding %d to fd_set", fd);
+	nfds = nfds > fd ? nfds : fd +1;
+	FD_SET(fd, &selects);
+}
+
+void clear_fd_select(int fd) {
+	FD_CLR(fd, &selects);
+}
+
+bool exiting = false;
+int selfpipe_write_end;
+static int libnet_client_main(const char *address, const char* service) {
+	struct addrinfo hints;
+	struct addrinfo *res;
+	struct addrinfo *res_head = 0;
+
+	memset(&hints, 0, sizeof hints);
+
+	hints.ai_family = AF_UNSPEC;
+	hints.ai_socktype = SOCK_STREAM;
+	hints.ai_protocol = IPPROTO_TCP;
+
+	int n;
+
+	log_info("Connecting to %s on %s", address, service);
+
+	check(!(n = getaddrinfo(address, service, &hints, &res)),
+			"getaddrinfo err for %s:%s: %s", address, service, gai_strerror(n));
+	res_head = res;
+
+	int selfpipe_fd;
+	check1((selfpipe_fd = create_selfpipe(&selfpipe_write_end, 0)), "selfpipe init");
+	do {
+		int sockfd;
+		if((sockfd = socket(res->ai_family, res->ai_socktype, res->ai_protocol) < 0)) {
+			continue;
+		}
+		add_fd_to_select(selfpipe_fd);
+		add_fd_to_select(sockfd);
+
+		if(!connect(sockfd, res->ai_addr, res->ai_addrlen)) {
+			log_err1("connect");
+			close_socket(sockfd);
+			continue;
+		}
+
+		while(!exiting)
+			handle_client_input(sockfd);
+
+		//TODO shit
+		close(sockfd);
+	} while (!(res = res->ai_next));
+
+	return 0;
+error:
+	if(res_head)
+		freeaddrinfo(res_head);
+
+	return 0;
+}
+
+static void* libnet_main_pthread_wrapper(void * args) {
+	const char **params = ((const char**)args);
+	//TODO(florek) return value!
+	const char *address = params[0];
+	const char *service = params[1];
+
+	libnet_client_main(address, service);
+	return 0;
+}
+
+pthread_t libnet_thread;
+
+bool libnet_thread_start(char *address, char *service) {
+	char **params = malloc(2 * sizeof * params);
+	params[0] = address;
+	params[1] = service;
+	log_info("XY : %s, %s", params[0], params[1]);
+	check1(!pthread_create(&libnet_thread, 0, libnet_main_pthread_wrapper, params),
+			"pthread_create libnet_thread");
+	return true;
+error:
+	return false;
+}
+
+static bool libnet_thread_shutdown() {
+	exiting = true;
+	check1(notify_selfpipe(selfpipe_write_end) >= 0, "selfpipe write");
+	//TODO(florek) error handling
+	//TODO(florek) return val hadnling
+	check1(!pthread_join(libnet_thread, 0), "pthread_join libnet_thread");
+	return true;
+error:
+	return false;
+}
+
+
+char addr[] = "localhost";
+char serv[] = "4200";
+
+int main(int argc, char **argv) {
+	log_info1("Starting client thread");
+	libnet_thread_start(addr, serv);
+	getchar();
+	log_info1("Shutting down server thread");
+	libnet_thread_shutdown();
+	return 0;
+}
+
