+++ libnet/common.h
@@ -1,0 +1,97 @@
+
+#include <semaphore.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include <sys/types.h>
+
+#define HEADER_LEN 2
+typedef struct {
+	int fd;
+	short state;
+	int read_status;
+	int retries;
+	unsigned char internal_buffer[HEADER_LEN]; // used to store partially retreived data
+} client_info;
+
+typedef struct {
+	unsigned char tag;
+	sem_t number;
+} available_tag_t;
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic warning "-pedantic"
+#pragma message "This is not strictly C99 but works and is awfully convenient"
+typedef struct {
+	unsigned char tag;
+	size_t length;
+	unsigned char value[0]; // variable size array at the tail of struct
+} tlv;
+#pragma GCC diagnostic pop
+
+// as per https://lists.gnu.org/archive/html/bug-gnulib/2006-06/msg00014.html
+#ifndef SSIZE_MAX
+# define SSIZE_MAX ((ssize_t) (SIZE_MAX /2))
+#endif
+
+#define TAG_INTERNAL_MASK 0xe0
+
+#define TAG_HELO 0xe0
+#define TAG_BYE  0xe1
+#define TAG_ECHO 0xe2
+#define TAG_OK   0xe3
+
+#define STATE_NONE 0 // free slot
+#define STATE_CONNECTING 1
+#define STATE_READY 2
+#define STATE_DISCONNECTED 3
+#define STATE_ERROR 4
+
+#define READ_STATUS_DEFAULT 0
+#define READ_STATUS_WAITING_FOR_DATA 1
+
+#ifndef MAX_MESSAGE_QUEUE
+# define MAX_MESSAGE_QUEUE 100
+#endif
+
+tlv** get_message_queue();
+
+bool handle_message(client_info *client, tlv *message) __attribute__((warn_unused_result));
+
+bool send_tag(client_info const * client, unsigned char tag,
+		size_t length, unsigned char * value)
+	__attribute__((warn_unused_result));
+
+//bool handle_internal_message(client_info *client, tlv *message);
+
+bool try_read(client_info *client, unsigned char *buff, size_t bytes_to_read);
+bool try_read_header(client_info *client, unsigned char *buff);
+
+bool close_socket(int fd);
+
+int handle_client_input(int sock_fd);
+
+bool wait_for_tag(unsigned char tag);
+
+bool enqueue_message(tlv *message);
+
+int create_selfpipe(int *selfpipe_write_end, int *selfpipe_read_end);
+int notify_selfpipe(int selfpipe_write_end);
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic warning "-Wredundant-decls"
+#pragma message "Yeah I know"
+bool libnet_init(unsigned char *tags_to_register, unsigned tags_to_register_number);
+#pragma GCC diagnostic pop
+
+#ifdef _DEBUG
+char const * state_to_string(int status);
+char const * read_status_to_string(int status);
+#endif
+
+
+// These functions need to be defined to be used as callbacks
+void clear_fd_select(int fd);
+
+unsigned get_client_id(client_info *client);
+client_info* get_client_by_fd(int fd);
+available_tag_t* get_available_tags_struct(unsigned char tag);
